+++ Src/Web/Web.Shared.Net.Tests/RequestTrackingTelemetryModuleTest.cs
@@ -15,12 +15,14 @@
     using Microsoft.ApplicationInsights.Extensibility.Implementation;
     using Microsoft.ApplicationInsights.Extensibility.W3C;
     using Microsoft.ApplicationInsights.TestFramework;
+    using Microsoft.ApplicationInsights.Web.Extensibility.Implementation;
     using Microsoft.ApplicationInsights.Web.Helpers;
     using Microsoft.ApplicationInsights.Web.Implementation;
     using Microsoft.ApplicationInsights.Web.TestFramework;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
-
+    using Xunit.Sdk;
     using Assert = Xunit.Assert;
+    using MsAssert = Microsoft.VisualStudio.TestTools.UnitTesting.Assert;
 
     /// <summary>
     /// Platform independent tests for RequestTrackingTelemetryModule.
@@ -476,6 +478,53 @@
             Assert.Equal(restoredActivity.StartTimeUtc, intermediateRequest.Timestamp);
             Assert.Equal(restoredActivity.Duration, intermediateRequest.Duration);
             Assert.True(intermediateRequest.Properties.ContainsKey("AI internal"));
+        }
+
+        [TestMethod]
+        public void VerifyBehaviorWhenDeferredIsFalse()
+        {
+            var config = this.CreateDefaultConfig(HttpModuleHelper.GetFakeHttpContext());
+            var context = HttpModuleHelper.GetFakeHttpContext();
+
+            // Create and Validate RequestTrackingTelemetryModule
+            var module = this.RequestTrackingTelemetryModuleFactory(config);
+            MsAssert.IsFalse(module.DisableTrackingProperties, $"{nameof(module.DisableTrackingProperties)} should be False by default.");
+
+            // Validate Telemetry Processor Chain
+            MsAssert.IsFalse(config.DefaultTelemetrySink.TelemetryProcessors.Any(x => x is PostSamplingTelemetryProcessor));
+
+            // Run test to generate requestTelemetry
+            module.OnBeginRequest(context);
+            module.OnEndRequest(context);
+            var requestTelemetry = context.GetRequestTelemetry();
+
+            // Validate requestTelemetry
+            MsAssert.IsNotNull(requestTelemetry, "TEST ERROR: Failed to create requestTelemetry.");
+            MsAssert.IsNotNull(requestTelemetry.Url);
+        }
+
+        [TestMethod]
+        public void VerifyBehaviorWhenDeferredIsTrue()
+        {
+            var context = HttpModuleHelper.GetFakeHttpContext();
+            var config = new TelemetryConfiguration();
+            config.ExperimentalFeatures.Add("deferRequestTrackingProperties");
+
+            // Create and Validate RequestTrackingTelemetryModule
+            var module = this.RequestTrackingTelemetryModuleFactory(config);
+            MsAssert.IsTrue(module.DisableTrackingProperties, "TEST ERROR: Module was not initialized with custom config.");
+
+            // Validate Telemetry Processor Chain
+            MsAssert.IsTrue(config.DefaultTelemetrySink.TelemetryProcessors.Any(x => x is PostSamplingTelemetryProcessor), "Module should inject PostSamplingTelemetryProcessor");
+
+            // Run test to generate requestTelemetry
+            module.OnBeginRequest(context);
+            module.OnEndRequest(context);
+            var requestTelemetry = context.GetRequestTelemetry();
+
+            // Validate requestTelemetry
+            MsAssert.IsNotNull(requestTelemetry, "TEST ERROR: Failed to create requestTelemetry.");
+            MsAssert.IsNotNull(requestTelemetry.Url); // set by PostSamplingTelemetryProcessor
         }
 
         private TelemetryConfiguration CreateDefaultConfig(HttpContext fakeContext, string rootIdHeaderName = null, string parentIdHeaderName = null, string instrumentationKey = null)